Ongoing alsa_capture refactoring
The objective is to get rid of the nasty signals. I feel like it is not
a good practive especially when glc is meant to run with host programs
that we do not have any control over.

Beside, with having a dedicated thread to capture the pcm stream, I do
not see any obvious benefits of using async notifications.