Experimental video filter.
The idea is to demux stream types
as soon as possible to avoid having audio
packets being copied and processed by 3 layers
of video filters.

The drawback of the idea, is that it introduce a new
buffer object where video frames will be copied to/from
buffer.

Since video data is probably much greater than 3 times
the amount of audio data, the price is probably higher
than the received gains.

Just keeping the code around since it could be useful
for other things eventually.