Allow the capture to be stopped from the sink
I found out that while with the file sink, it was quite exceptional
to have an error occuring in the sink while capturing, it is quite
easy to see with the pipe sink. The most common errors are:

1. While experimenting with the external program, it just abort
   immediatly because of an invalid combination of cmdline params
2. It is too slow to consume the stream at the pace we create it.

Prior this commit, the sink thread was exiting and the only way to
restart the capture  was to restart the host app.

Now the sink has a callback to stop the capture which will
drain the content of the sink buffer and leave the whole thing
in a consistent state so capture will start correctly at the
next attempt.

Signed-off-by: Olivier Langlois <olivier@trillion01.com>